+++ src/main/java/duke/Parser.java
@@ -1,3 +1,4 @@
+package duke;
 import java.util.Scanner;
 
 class Parser {
