+++ src/main/java/duke/Task.java
@@ -1,0 +1,24 @@
+package duke;
+
+class Task {
+    protected String description;
+    protected boolean isDone;
+
+    public Task(String description) {
+        this.description = description;
+        this.isDone = false;
+    }
+
+    public String getStatusIcon() {
+        return (isDone ? "X" : " "); // mark done task with X
+    }
+
+    public void markAsDone(){
+        this.isDone = true;
+    }
+
+    @Override
+    public String toString(){
+        return "[" + getStatusIcon() + "] " + description;
+    }
+}
