+++ src/main/java/duke/ToDo.java
@@ -1,0 +1,14 @@
+package duke;
+
+class ToDo extends Task {//No deadline
+
+
+    public ToDo(String description) {
+        super(description);
+    }
+
+    @Override
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+}
